Reverted `load_state_dict` patching: This didn't work because `load_state_dict` isn't called on child modules Created our own `load_state_dict` wrapper util that ignores scale factor keys
