+++ tests/models/test_gemnet_oc.py
@@ -18,7 +18,7 @@ from ase.io import read
 
 from ocpmodels.common.registry import registry
 from ocpmodels.common.transforms import RandomRotate
-from ocpmodels.common.utils import setup_imports
+from ocpmodels.common.utils import load_state_dict, setup_imports
 from ocpmodels.datasets import data_list_collater
 from ocpmodels.preprocessing import AtomsToGraphs
 
@@ -112,7 +112,7 @@ def load_model(request):
     new_dict = {
         k[len("module.") * 2 :]: v for k, v in checkpoint["state_dict"].items()
     }
-    model.load_state_dict(new_dict)
+    load_state_dict(model, new_dict)
 
     request.cls.model = model
